vista resultado e informacion, consumo de servicio de configuracion y soporte para decimales

+++ app/src/main/java/com/example/calculadoraapp/models/Material.kt
@@ -1,3 +1,3 @@
 package com.example.calculadoraapp.models
 
-data class Material(var material: String = "", var cantidad: Int, var unidadMedida: String = "")
+data class Material(var material: String = "", var cantidad: Double, var unidadMedida: String = "")
